+++ CHANGELOG.md
@@ -22,6 +22,11 @@
 * Changed `getJson` to have values be null if they aren't found rather than an empty string.
 * Implemented the `getJson` method correctly for a number of classes.
 * Changed `Task.percentComplete` to always return a float.
+* Changed the `NotImplementedError` for custom attachment handler not being found to `FeatureNotImplemented`. Additionally, changed the error message to specify the CLSID found on the attachment to better enable people to report issues.
+* Changed code for `Recipient` and `MessageBase` that makes it rely on `MessageBase.recipientTypeClass` to determine the class to use for the `recipientType` property. Adjusted the typing of `Recipient` to have it reflect the type that will be used.
+* Correctly changed the returned value for `ResponseStatus.fromIter` to actually return a List instead of a set.
+* Filled out typing information for a significant portion of the module where variables or functions were missing it.
+* Corrected a number of minor issues.
 
 **v0.45.0**
 * BREAKING: Changed parsing of string multiple properties to remove the trailing null byte. This *will* cause the output of parsing them to differ.